backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
opencv-python>=4.9.0
numpy>=1.26.0
openai-whisper>=20231117
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
httpx>=0.26.0
python-dotenv>=1.0.0
loguru>=0.7.2
